move get_member_data to be a utility funciton
