+++ apps/member/views.py
@@ -1,12 +1,11 @@
 import requests
 
-from django.conf import settings
 from django.contrib.auth.decorators import login_required
 from django.contrib.auth.mixins import LoginRequiredMixin
-from django.shortcuts import get_object_or_404, Http404, render, redirect, reverse
+from django.shortcuts import get_object_or_404, redirect, reverse
 from django.views.decorators.http import require_POST
 from django.views.generic.base import TemplateView
-from django.urls import reverse, reverse_lazy
+from django.urls import reverse_lazy
 from django.views.generic.edit import CreateView, DeleteView, UpdateView
 from django.views.generic.detail import DetailView
 
@@ -155,28 +154,3 @@ def revoke_resource_request(request, pk):
     if getattr(resource_request, 'resourcegrant', None):
         resource_request.resourcegrant.delete()
     return redirect(reverse('member:dashboard'))
-
-
-@login_required(login_url='home')
-def get_member_data(request, pk, resource_name, record_type):
-    # Get the path to the resource class from the settings, based on the resource_name.
-    resource_class_path = settings.RESOURCE_NAME_AND_CLASS_MAPPING.get(resource_name, None)
-    # If there is not a path for the resource_name, raise an error
-    if not resource_class_path:
-        raise Http404
-
-    # Is the record_type is not valid, raise an error
-    if record_type not in settings.VALID_MEMBER_DATA_RECORD_TYPES:
-        raise Http404
-
-    # Does the request.user's Organization have access to this member's resource?
-    resource_grant = get_object_or_404(
-        ResourceGrant.objects.filter(
-            member_id=pk,
-            resource_class_path=resource_class_path,
-            organization__users=request.user
-        )
-    )
-
-    data = resource_grant.resource_class(resource_grant.member).get()
-    return render(request, 'member/data.html', context={'data': data})
